+++ example/windows/flutter/generated_plugins.cmake
@@ -7,7 +7,6 @@ list(APPEND FLUTTER_PLUGIN_LIST
   firebase_core
   flutter_secure_storage_windows
   local_auth_windows
-  sentry_flutter
   supa_architecture
   url_launcher_windows
 )