+++ lib/supa_architecture_method_channel.dart
@@ -37,11 +37,9 @@ class MethodChannelSupaArchitecture extends SupaArchitecturePlatform {
   @override
   Future<void> initialize({
     bool useFirebase = false,
-    bool useSentry = false,
   }) async {
     super.initialize(
       useFirebase: useFirebase,
-      useSentry: useSentry,
     );
     cookieStorage = await HiveCookieManager.create();
     secureStorage.initialize();